EVA-1278 Map variants to java model

+++ eva-accession-import/src/main/java/uk/ac/ebi/eva/accession/dbsnp/processors/AssemblyCheckerProcessor.java
@@ -21,7 +21,7 @@ public class AssemblyCheckerProcessor implements ItemProcessor<SubSnpNoHgvs, Sub
     @Override
     public SubSnpNoHgvs process(SubSnpNoHgvs subSnpNoHgvs) throws Exception {
         ContigSynonyms contigSynonyms;
-        int start;
+        long start;
         if (subSnpNoHgvs.getChromosome() != null) {
             contigSynonyms = contigMapping.getContigSynonyms(subSnpNoHgvs.getChromosome());
             start = subSnpNoHgvs.getChromosomeStart();
